int cast removed

+++ mocpy/utils.py
@@ -83,7 +83,7 @@ def uniq2orderipix(uniq):
     to a (norder, ipix) tuple
     """
     order = ((np.log2(uniq//4)) // 2)
-    order = order.astype(int)
+    #order = order.astype(int)
     ipix = uniq - 4 * (4**order)
 
     return order, ipix
